--- api/test.js
@@ -1,3 +1,0 @@
-export default function(_req, res) {
-    res.status(200).send(process.env.TEST_ENV_VAR)
-}
