+++ package.json
@@ -2,12 +2,10 @@
   "name": "sorcsap.app-backend",
   "version": "1.0.0",
   "description": "",
-  "type": "module",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon app.js --config config.json",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "serve": "vercel dev"
+    "start": "vercel dev --listen 8000",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
